test(ir): isolate multi-source cutover evidence
Move the #4589 route, artifact, and runtime assertions into its dedicated focused suite and restore the inherited #2138 file exactly.

This keeps the canonical evidence while avoiding a pre-existing Node 25 512 MiB changed-root OOM; the restored #2138 suite still passes with its established heap requirement.
